User experience for Reading Lists, part 1.
- Implements the UI for viewing and managing Reading Lists, accessible
from the Nav menu.

- Implements the UI for adding an article to a reading list, by
pressing the "bookmark" icon below the lead image.

The business logic of dealing with the actual collection of reading lists
is handled by an interface called ReadingListData. In this patch, the
interface is implemented by a static hard-coded list. In the final
implementation, it should be handled by actual DB interactions.
Additionally, in the final implementation, certain elements will probably
need to be converted to asynchronous operation.

Note: the "Reading Lists" nav menu item will coexist with the "Saved
Pages" item, and is feature-flagged for non-production builds. Saved Pages
will be completely removed when reading lists are fully implemented.

Bug: T127711
Bug: T129590

+++ app/src/main/java/org/wikipedia/page/leadimages/LeadImagesHandler.java
@@ -20,16 +20,19 @@ import org.json.JSONObject;
 import org.wikipedia.R;
 import org.wikipedia.analytics.GalleryFunnel;
 import org.wikipedia.page.Page;
+import org.wikipedia.page.PageActivity;
 import org.wikipedia.page.PageTitle;
 import org.wikipedia.WikipediaApp;
 import org.wikipedia.bridge.CommunicationBridge;
 import org.wikipedia.page.PageFragment;
 import org.wikipedia.page.gallery.GalleryActivity;
+import org.wikipedia.readinglist.AddToReadingListDialog;
 import org.wikipedia.savedpages.DeleteSavedPageTask;
 import org.wikipedia.savedpages.SavePageTask;
 import org.wikipedia.savedpages.SavedPage;
 import org.wikipedia.util.DimenUtil;
 import org.wikipedia.util.FeedbackUtil;
+import org.wikipedia.util.ReleaseUtil;
 import org.wikipedia.util.ShareUtil;
 import org.wikipedia.util.UriUtil;
 import org.wikipedia.views.ObservableWebView;
@@ -346,11 +349,15 @@ public class LeadImagesHandler {
             if (getPage() == null) {
                 return;
             }
-
-            if (bookmarkSaved) {
-                saveBookmark();
+            // TODO: resolve when Saved Pages is removed
+            if (ReleaseUtil.isProdRelease()) {
+                if (bookmarkSaved) {
+                    saveBookmark();
+                } else {
+                    deleteBookmark();
+                }
             } else {
-                deleteBookmark();
+                ((PageActivity) getActivity()).showBottomSheet(AddToReadingListDialog.newInstance(getTitle()));
             }
         }
